+++ tests/test-splay-tree.c
@@ -23,8 +23,8 @@
 #include <CUnit/Basic.h>
 #include <float.h>
 #include <stdarg.h>
-#include "cdcontainers/splay-tree.h"
-#include "cdcontainers/casts.h"
+#include <cdcontainers/splay-tree.h>
+#include <cdcontainers/casts.h>
 
 static struct cdc_pair a = {CDC_INT_TO_PTR(0), CDC_INT_TO_PTR(0)};
 static struct cdc_pair b = {CDC_INT_TO_PTR(1), CDC_INT_TO_PTR(1)};
@@ -43,7 +43,7 @@ static int lt_int(const void *l, const void *r)
 static bool splay_tree_key_int_eq(struct cdc_splay_tree *t, size_t count, ...)
 {
         va_list args;
-        int i;
+        size_t i;
         struct cdc_pair *val;
         void *tmp;
 
@@ -188,7 +188,7 @@ void test_splay_tree_clear()
 void test_splay_tree_insert()
 {
         struct cdc_splay_tree *t;
-        int i, count = 100;
+        size_t i, count = 100;
         bool failed = false;
         void *val;
 
@@ -314,7 +314,7 @@ void test_splay_tree_iterators()
         struct cdc_splay_tree_iter it1, it2;
         struct cdc_pair *arr[] = {&a, &b, &c, &d, &e, &f, &g, &h};
         bool check;
-        int i;
+        size_t i;
 
         CU_ASSERT(cdc_splay_tree_ctorl1(&t, NULL, lt_int,
                                         &a, &b, &c, &d, &e, &f, &g, &h, NULL) == CDC_STATUS_OK);
